refactor: Share stream workspace alias validation
