+++ internal/service/streamconnection/workspace_name_plan_modifier_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/stretchr/testify/assert"
 
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/streamalias"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/streamconnection"
 )
 
@@ -46,7 +47,7 @@ func TestIsWorkspaceNameAliasReversion(t *testing.T) {
 
 	for name, tc := range testCases {
 		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, tc.reverted, streamconnection.IsWorkspaceNameAliasReversion(tc.stateWorkspaceName, tc.stateInstanceName, tc.planWorkspaceName, tc.planInstanceName))
+			assert.Equal(t, tc.reverted, streamalias.IsWorkspaceNameAliasReversion(tc.stateWorkspaceName, tc.stateInstanceName, tc.planWorkspaceName, tc.planInstanceName))
 		})
 	}
 }
