+++ internal/service/streamprocessor/model.go
@@ -20,17 +20,6 @@ func GetWorkspaceOrInstanceName(workspaceName, instanceName types.String) string
 	return instanceName.ValueString()
 }
 
-// IsWorkspaceNameAliasReversion reports whether a resource with canonical
-// workspace_name-only state is being changed back to deprecated instance_name.
-func IsWorkspaceNameAliasReversion(stateWorkspaceName, stateInstanceName, planWorkspaceName, planInstanceName types.String) bool {
-	// Legacy imports populated both aliases. Only workspace_name-only state is
-	// canonical and therefore subject to the one-way migration restriction.
-	return !stateWorkspaceName.IsNull() && !stateWorkspaceName.IsUnknown() &&
-		stateInstanceName.IsNull() &&
-		planWorkspaceName.IsNull() &&
-		!planInstanceName.IsNull() && !planInstanceName.IsUnknown()
-}
-
 // IsAliasOnlyTransition reports whether a plan preserves the effective workspace
 // and changes no processor configuration other than its alias representation.
 func IsAliasOnlyTransition(plan, state *TFStreamProcessorRSModel) bool {
